Move `SoftContactsState` to `soft_contacts`
